great things

+++ app/parse.rb
@@ -26,7 +26,7 @@ module Forked
         # Elements understood by the parser:
         # [x] :blockquote (physical div)
         # :block (logical div)
-        # :action (single line code)
+        # CANCELLED :action (single line code)
         # :action_block (multiline code)
         # :code (present with code format < 1 line)
         # [x] :code_block (present code style, multiline)
@@ -71,12 +71,10 @@ module Forked
         story_file.each_line.with_index do |line, line_no|
           # putz "# #{line_no}: #{line.strip}" 
 
-          ### PRESERVE LINE (^%)
+          ### PRESERVE LINE (^%) and stop parsing it
           result = parse_preserve_line(line, context)
-
-          unless context.intersection([:title, :block, :long_code]).any?
-            if (result = parse_preserver(line))
-              story[:chunks][-1][:content] << {
+          if result
+            story[:chunks][-1][:content] << {
                 type: :paragraph,
                 atoms: [
                   {
@@ -86,35 +84,20 @@ module Forked
                 ]
               }
               next
-            end
           end
 
-          ### STRIP COMMENT (//)
-          unless context.intersection([:code_block]).any?
-            result = parse_comment(line)
+          ### STRIP COMMENT (//) and continue parsing line
+            result = parse_comment(line, context)
             line = result if result
-          end
 
           ### TITLE
-          if (result = parse_title(line))
-            if (context == [:title]) && result && !result.empty? && story[:title] == DEFAULT_TITLE
-              story[:title] = result
-              context.delete(:title) # title is found
-              context << :heading  # second element must be a heading
-              # no further processing on this line
-              next
-            end
-          end
-
-          # exception if title is not first line
-          if context == [:title] && !line.strip.empty?
-            raise "FORKED: CONTENT BEFORE TITLE.
-
-The first line of the story file is expected to be the title.
-Please add a title to the top of the Story File. Example:
-
-`# The Name of this Story`
-"
+          result = parse_title(line, context)
+          if result 
+            story[:title] = result
+            context.delete(:title) # title is found
+            context << :heading  # second element must be a heading
+            # no further processing on this line
+            next
           end
 
           ### BLOCKQUOTE
@@ -131,17 +114,20 @@ Please add a title to the top of the Story File. Example:
 
           ### HEADING LINE
 
-          unless (context.intersection([:code_block])).any?
-            result = parse_heading(line)
+          # unless (context.intersection([:code_block])).any?
+            result = parse_heading(line, context)
             if result
               heading, chunk_id = result
-
+              putz result
               if heading.empty? && chunk_id.nil?
                 raise "Forked: Expected heading and/or ID on line #{line_no + 1}."
               end
 
+              heading = story.title if heading.empty?
+
               story[:chunks] << {
                 id: chunk_id,
+                actions: [],
                 content: [
                   {
                     type: :heading,
@@ -155,11 +141,11 @@ Please add a title to the top of the Story File. Example:
                 ]
               }
 
-              context.delete(:heading)
+              # context.delete(:heading)
 
               next
             end
-          end
+          # end
 
           if context.include?(:heading) && !line.strip.empty?
             # we were looking for a heading and didn't find one
@@ -187,24 +173,12 @@ Please add a heading line after the title and before any other content. Example:
           end
 
           ### CODE BLOCK
+          result = parse_code_block(line, context, story)
+          next if result
 
-          if (result = parse_code_block(line))
-            if !context.include?(:code_block)
-              context << (:code_block)
-              story[:chunks][-1][:content] << {
-                type: :code_block,
-                text: ''
-              }
-            else
-              context.delete(:code_block)
-            end
-            next
-          end
-
-          if context.include?(:code_block)
-            story[:chunks][-1][:content][-1][:text] += line
-            next
-          end
+          ### ACTION BLOCK
+          result = parse_action_block(line, context, story)
+          next if result
 
           # PARAGRAPH
           unless context.intersection([:title]).any?
@@ -278,27 +252,30 @@ Please add a heading line after the title and before any other content. Example:
       # check to see if it is safe to proceed based on
       # context rules
       def context_safe?(context, prohibited, mandatory)
-        array_intersect?(context, prohibited) &&
-        array_intersect?(context, mandatory)
+        context_prohibited = array_intersect?(context, prohibited)
+        context_mandatory = mandatory.empty? ||  array_intersect?(context, mandatory) 
+
+        # putz "context_prohibited #{context_prohibited}"
+        # putz "context_mandatory  #{context_mandatory}"
+
+        context_mandatory && !context_prohibited
       end
 
+      # true if arr1 and arr2 contain any overlapping elements
       def array_intersect?(arr1, arr2)
         arr1.intersection(arr2).any?
       end
 
+      # Force a line to display as plain, unformatted text, ignoring any further markup
+      # Used for troubleshooting, not part of the specification
       def parse_preserve_line(line, context)
-        # forbidden contexts
+        return unless line.strip.start_with?('%')
+
         prohibited_contexts = [:title]
-        mandatory_contexts = [:title]
-
+        mandatory_contexts = []
         return unless context_safe?(context, prohibited_contexts, mandatory_contexts)
-
-        return unless line.strip.start_with?('%') 
-
-
-          line.delete_prefix('%')
-
-        return line
+        
+        line.delete_prefix('%')
       end
 
       def parse_paragraph(line)
@@ -307,20 +284,13 @@ Please add a heading line after the title and before any other content. Example:
         line.split('\n')
       end
 
-      def parse_preserver(line)
-        # if line starts with #
-        # don't parse it, just treat it as unstyled text and move on
-        # Real line start. Doesn't have any effect inside blocks.
-        
-        if line.strip.start_with?('%')
-          line.delete_prefix('%')
-        end
-      end
+      # strip comments from line (comments begin with //)
+      def parse_comment(line, context)
+        return unless line.include?('//')
 
-      def parse_comment(line)
-        # strip out comments from line
-        # support only single line comments for now
-        # add multiline comments to context in future
+        prohibited_contexts = [:code_block]
+        mandatory_contexts = []
+        return unless context_safe?(context, prohibited_contexts, mandatory_contexts)
 
         if line.include?(' //')
           index = line.index(' //')
@@ -330,7 +300,6 @@ Please add a heading line after the title and before any other content. Example:
         end
 
         nil
-        # line
       end
 
       def parse_blockquote(line)
@@ -344,57 +313,65 @@ Please add a heading line after the title and before any other content. Example:
         end
       end
 
-      def parse_code_block(line)
-        return unless line.include?('```')
-        # Code format blocks are surrounded with  ``` ```
-        # The first line of the example should begin with three ticks
-        # The last line of the example should end with three ticks
-        
-        # code format blocks look like code but don't get executed
-        # They are likely to contain backticks which should be ignored
-        # and processed as normal text.
-         line.include? '```'
-        
-        # line.strip.split('```')
-
-        # position = nil
-        # if line.strip.start_with?('```')
-        #   line.lstrip!
-        #   line.delete_prefix!('```')
-        #   position = :start
-        # end
-
-        # if line.strip.end_with?('```')
-        #   line.rstrip!
-        #   line.delete_suffix!('```')
-        #   position = position == :start ? :both : :end
-        # end
-
-        # if position
-        #   [position, line]
-        # end
-      end
-
-      def parse_action(line)
-        # actions are surrounded with ` ` for one line code
-        # or with ``` ``` for multiline code
-
-        # if we meet an opening ``` with no closure in the same line,
-        # open a context and keep reading lines until the ``` appears.
-
-        # if no ``` appears by the end of the chunk, terminate it or
-        # raise an error.
-
-        # DON'T catch it if:
-        # line starts with %
-        # escaped (\```)
-        # Inside a code formatted block (~~~)
-
-        if line.include?('```')
-          # # line
+      # Code blocks format code for display
+      # They begin with three ticks ``` on a blank line
+      # and end with three ticks on a blank line
+      def parse_code_block(line, context, story)
+        if line.include? '```'
+          if context.include?(:code_block)
+            context.delete(:code_block)
+          else
+            context << (:code_block)
+            story[:chunks][-1][:content] << {
+              type: :code_block,
+              text: ''
+            }
+          end
+          true
+        elsif context.include?(:code_block)
+          story[:chunks][-1][:content][-1].text += line
+          true
         end
-
       end
+
+      # Code blocks format code for display
+      # They begin with three ticks ``` on a blank line
+      # and end with three ticks on a blank line
+      def parse_action_block(line, context, story)
+        if line.include? '^^^'
+          if context.include?(:action_block)
+            context.delete(:action_block)
+          else
+            context << (:action_block)
+            story[:chunks][-1][:actions] << ''
+          end
+          true
+        elsif context.include?(:action_block)
+          story[:chunks][-1][:actions][-1] += line
+          true
+        end
+      end
+
+      # def parse_action(line)
+      #   # actions are surrounded with ` ` for one line code
+      #   # or with ``` ``` for multiline code
+
+      #   # if we meet an opening ``` with no closure in the same line,
+      #   # open a context and keep reading lines until the ``` appears.
+
+      #   # if no ``` appears by the end of the chunk, terminate it or
+      #   # raise an error.
+
+      #   # DON'T catch it if:
+      #   # line starts with %
+      #   # escaped (\```)
+      #   # Inside a code formatted block (~~~)
+
+      #   if line.include?('```')
+      #     # # line
+      #   end
+
+      # end
 
       def parse_block(line)
         # blocks start with <
@@ -406,41 +383,48 @@ Please add a heading line after the title and before any other content. Example:
 
       end
 
-      def parse_title(line)
-          # titles must BEGIN with a single # (not double)
-          # if a title is not given, a default will be used.
-          # if there is more than one title, only the first one is used.
+      # The title is required. No content can come before it.
+      # The Title line starts with a single #
+      def parse_title(line, context)
+        prohibited_contexts = []
+        mandatory_contexts = [:title]
+        return unless context_safe?(context, prohibited_contexts, mandatory_contexts)
 
-          # The title is assumed to be the entire line
-          # No styles or other formatting can be applied to the title.
+        if line.strip.start_with?('#') && !line.strip.start_with?('##')
+          line.strip.delete_prefix('#').strip
+        elsif !line.strip.empty?
+          raise "FORKED: CONTENT BEFORE TITLE.
 
-          # future: If the user includes a level 1 heading in a chunk
-          # It's formatting. Display as a heading.
+The first line of the story file is expected to be the title.
+Please add a title to the top of the Story File. Example:
 
-          if line.strip.start_with?('#') && !line.strip.start_with?('##')
-            line.strip.delete_prefix('#').strip
+`# The Name of this Story`
+"
           end
       end
 
-      def parse_heading(line)
-        # the heading line must begin with a double #
-        # It can have:
-        # Text only
-        # an ID only
-        # Text and ID
+      # The heading line starts a new chunk
+      # The heading line begins with a double #
+      def parse_heading(line, context)
+        return unless line.strip.start_with?('##') && !line.strip.start_with?('###')
 
-        # new heading breaks all contexts. Any open contexts should either 
-        # be forced closed or errored.
+        prohibited_contexts = [:code_block]
+        mandatory_contexts = []
+        return unless context_safe?(context, prohibited_contexts, mandatory_contexts)
 
-        if line.strip.start_with?('##') && !line.strip.start_with?('###')
-          line.strip!.delete_prefix!('##').strip!
+        context.delete(:heading)
+        # raise "CONTEXT ERROR. Context should be empty but context is #{context}" unless context.empty?
 
-          if line.include?('{') && line.include?('}')
-            line = pull_out('{', '}', line)
-          else
-            [line]
-          end
+        line.strip!
+        line.delete_prefix!('##').strip!
+
+        if line.include?('{') && line.include?('}')
+          line = pull_out('{', '}', line)
+        else
+          line = [line]
         end
+
+        line
       end
 
       def parse_trigger(line)
